pyproject.toml
[project]
name = "spotexample"
version = "0.1.0"
description = "A ocelescope plugin to test spot"
readme = "README.md"
authors = [
    { name = "Grkmr", email = "goerkem@posteo.com" },
]
maintainers = [
    { name = "Grkmr", email = "goerkem@posteo.com" },
]
license = { text = "MIT" }

dependencies = [
    "ocelescope~=0.1.3",
    "spot-local",
]
requires-python = ">=3.13"

[dependency-groups]
dev = ["black>=25.1.0", "pre-commit>=4.2.0", "ruff>=0.12.7"]


[tool.ty]
# All rules are enabled as "error" by default; no need to specify unless overriding.
# Example override: relax a rule for the entire project (uncomment if needed).
# rules.TY015 = "warn"  # For invalid-argument-type, warn instead of error.

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # Pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "UP", # pyupgrade
]

[tool.uv]
package = true
environments = [
  "sys_platform == 'linux' and platform_machine == 'x86_64'",
  "sys_platform == 'linux' and platform_machine == 'aarch64'",
]

[tool.uv.sources]
spot-local = [
  { path = "src/spotexample/wheels/spot_local-2.14.5.dev0-py3-none-manylinux_2_34_x86_64.whl", marker = "sys_platform == 'linux' and platform_machine == 'x86_64'" },
  { path = "src/spotexample/wheels/spot_local-2.14.5.dev0-py3-none-manylinux_2_34_aarch64.whl", marker = "sys_platform == 'linux' and platform_machine == 'aarch64'" },
]

[build-system]
requires = ["uv_build>=0.8.4,<0.9.0"]
build-backend = "uv_build"
